requirements.txt
google-cloud-aiplatform
openai
python-dotenv
supabase
PyMuPDF
requests
sentence-transformers
